% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_search.R
\name{create_random_search}
\alias{create_random_search}
\title{Create a Random Search Space}
\usage{
create_random_search(nrounds = 100, params, seed = 1234)
}
\arguments{
\item{nrounds}{number of random iterations}

\item{params}{a list of parameters with low and high values}

\item{seed}{random seed}
}
\description{
Create a Random Search Space
}
