% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_search.R
\name{random_search}
\alias{random_search}
\title{Search along a random search grid}
\usage{
random_search(grid, FUN, maximise = FALSE, verbose = FALSE)
}
\arguments{
\item{grid}{random search grid}

\item{FUN}{function to be applied}

\item{maximise}{boolean value indicate weather parameters that maximise a function
should be selected. Defaults to FALSE}

\item{verbose}{boolean value that controls output verbosity. Defaults to FALSE.}
}
\description{
Search along a random search grid
}
