% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{Search along a random search grid}
\usage{
search(grid, FUN, maximise = FALSE)
}
\arguments{
\item{grid}{random search grid}

\item{FUN}{function to be applied}

\item{maximise}{boolean value indicate weather parameters that maximise a function
should be selected}
}
\description{
Search along a random search grid
}
